processor.py
# processor.py
import cv2
import pytesseract
from PIL import Image
from presidio_analyzer import AnalyzerEngine
from presidio_anonymizer import AnonymizerEngine
from langchain_community.llms import OpenAI # or GooglePalm, etc.
from langchain.chains import create_extraction_chain

# --- 1. IMAGE PRE-PROCESSING (OpenCV) ---
def clean_image_for_ocr(image_bytes):
    # Use OpenCV to read and clean the image
    # (e.g., convert to grayscale, thresholding)
    # This improves OCR accuracy.
    # ... your OpenCV logic here ...
    return 0

# --- 2. TEXT EXTRACTION (OCR / Parsers) ---
def extract_text_from_file(file):
    # Check file type
    if file.type in ["image/png", "image/jpeg"]:
        image = Image.open(file)
        # Optional: Clean the image first
        # cleaned_image = clean_image_for_ocr(image)
        raw_text = pytesseract.image_to_string(image)
    # Add logic for PDF, DOCX, etc.
    # elif file.type == "application/pdf":
    # ...
    else:
        raw_text = "File type not supported for text extraction yet."
    return raw_text

# --- 3. PII CLEANSING (Presidio) ---
def anonymize_text(text):
    analyzer = AnalyzerEngine()
    anonymizer = AnonymizerEngine()
    
    analyzer_results = analyzer.analyze(text=text, language='en')
    anonymized_result = anonymizer.anonymize(text=text, analyzer_results=analyzer_results)
    
    return anonymized_result.text

# --- 4. DATA EXTRACTION (LangChain) ---
def extract_insights(text):
    # Define a schema for the data you want to extract
    schema = {
        "properties": {
            "firewall_rule": {"type": "string"},
            "iam_policy_statement": {"type": "string"},
        }
    }
    # Initialize your LLM and the extraction chain
    llm = OpenAI(temperature=0) # Or your preferred LLM
    chain = create_extraction_chain(schema, llm)
    
    return chain.run(text)

# --- THE MAIN WORKFLOW FUNCTION ---
def process_document(uploaded_file):
    # Step 1 & 2: Get raw text from the file
    raw_text = extract_text_from_file(uploaded_file)
    
    # Step 3: Cleanse the text using Presidio
    clean_text = anonymize_text(raw_text)
    
    # Step 4: Extract structured data using LangChain
    extracted_data = extract_insights(clean_text)
    
    return clean_text, extracted_data