requirements.txt

acres                     
aiohappyeyeballs          
aiohttp                   
aiosignal                 1.
altair                    5.5.0
annotated-types           0.7.0
anyio                     4.10.0
attrs                     25.3.0
blinker                   1.9.0
blis                      1.3.0
cachetools                6.2.0
catalogue                 2.0.10
certifi                   2025.8.3
cffi                      1.17.1
charset-normalizer        3.4.3
ci-info                   0.3.0
click                     8.2.1
cloudpathlib              0.22.0
colorama                  0.4.6
confection                0.1.5
configobj                 5.0.9
configparser              7.2.0
cryptography              44.0.3
cymem                     2.0.11
dataclasses-json          0.6.7
distro                    1.9.0
en_core_web_lg            3.8.0
etelemetry                0.3.1
filelock                  3.19.1
fitz                      0.0.1.dev2
frozenlist                1.7.0
funcparserlib             1.0.1
gitdb                     4.0.12
GitPython                 3.1.45
greenlet                  3.2.4
groq                      0.31.1
h11                       0.16.0
httpcore                  1.0.9
httplib2                  0.31.0
httpx                     0.28.1
httpx-sse                 0.4.1
hy                        1.1.0
idna                      3.10
Jinja2                    3.1.6
jiter                     0.10.0
jsonpatch                 1.33
jsonpointer               3.0.0
jsonschema                4.25.1
jsonschema-specifications 2025.9.1
langchain                 0.3.27
langchain-community       0.3.29
langchain-core            0.3.75
langchain-groq            0.3.8
langchain-openai          0.3.32
langchain-text-splitters  0.3.11
langcodes                 3.5.0
langsmith                 0.4.27
language_data             1.3.0
looseversion              1.3.0
lxml                      6.0.1
marisa-trie               1.3.1
markdown-it-py            4.0.0
MarkupSafe                3.0.2
marshmallow               3.26.1
mdurl                     0.1.2
multidict                 6.6.4
murmurhash                1.0.13
mypy_extensions           1.1.0
narwhals                  2.4.0
networkx                  3.5
nibabel                   5.3.2
nipype                    1.10.0
numpy                     2.2.6
openai                    1.107.0
opencv-python             4.12.0.88
orjson                    3.11.3
packaging                 25.0
pandas                    2.3.2
pathlib                   1.0.1
phonenumbers              9.0.13
pillow                    11.3.0
pip                       25.2
preshed                   3.0.10
presidio_analyzer         2.2.359
presidio_anonymizer       2.2.359
processor                 0.10.0
propcache                 0.3.2
protobuf                  6.32.0
prov                      2.1.1
puremagic                 1.30
pyarrow                   21.0.0
pycparser                 2.22
pydantic                  2.11.7
pydantic_core             2.33.2
pydantic-settings         2.10.1
pydeck                    0.9.1
pydot                     4.0.1
Pygments                  2.19.2
PyMuPDF                   1.26.4
pyparsing                 3.2.4
pytesseract               0.3.13
python-dateutil           2.9.0.post0
python-dotenv             1.1.1
python-pptx               1.0.2
pytz                      2025.2
pyxnat                    1.6.3
PyYAML                    6.0.2
rdflib                    7.1.4
referencing               0.36.2
regex                     2025.9.1
requests                  2.32.5
requests-file             2.1.0
requests-toolbelt         1.0.0
rich                      14.1.0
rpds-py                   0.27.1
scipy                     1.16.2
setuptools                80.9.0
shellingham               1.5.4
simplejson                3.20.1
six                       1.17.0
smart_open                7.3.1
smmap                     5.0.2
sniffio                   1.3.1
spacy                     3.8.7
spacy-legacy              3.0.12
spacy-loggers             1.0.5
SQLAlchemy                2.0.43
srsly                     2.5.1
streamlit                 1.49.1
tenacity                  9.1.2
thinc                     8.3.6
tiktoken                  0.11.0
tldextract                5.3.0
toml                      0.10.2
tornado                   6.5.2
tqdm                      4.67.1
traits                    7.0.2
Twiggy                    0.4.7
twiggy-goodies            0.12.0
typer                     0.17.4
typing_extensions         4.15.0
typing-inspect            0.9.0
typing-inspection         0.4.1
tzdata                    2025.2
urllib3                   2.5.0
wasabi                    1.1.3
watchdog                  6.0.0
weasel                    0.4.1
wrapt                     1.17.3
xlsxwriter                
yarl                      
zstandard                 